% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{BuildFisheryTable}
\alias{BuildFisheryTable}
\title{Make a table of fishery mortality}
\usage{
BuildFisheryTable(fishery.mortality, column.name)
}
\arguments{
\item{fishery.mortality}{Escapement values from a FRAM run}

\item{column.name}{The list of PSC stocks}
}
\value{
dataframe of fishery mortalities
}
\description{
Make a table of fishery mortality
}
