% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{checkFramCommentCol}
\alias{checkFramCommentCol}
\title{Check if the FRAM database as the new Comments fields for importing, if not add
the columns to the appropriate tables}
\usage{
checkFramCommentCol(fram_db_conn)
}
\arguments{
\item{fram_db_conn}{An ODBC connection to the FRAM MS Access database}
}
\description{
Check if the FRAM database as the new Comments fields for importing, if not add
the columns to the appropriate tables
}
