% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{combine_GS_fishery_mortality}
\alias{combine_GS_fishery_mortality}
\title{Combines the GS MUs for the fishery.mortality DF (pre or post season)}
\usage{
combine_GS_fishery_mortality(fishery.mortality)
}
\arguments{
\item{stock.summary}{stock.summary from the pre or post season list}

\item{run.year}{inlcude.old.GS.MUs logical indicating if the ML and VI GS MUs should still be in the table}
}
\description{
Combines the GS MUs for the fishery.mortality DF (pre or post season)
}
