% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{combine_GS_stock_summary}
\alias{combine_GS_stock_summary}
\title{Combines the GS MUs for the stock.summary DF (pre or post season)}
\usage{
combine_GS_stock_summary(stock.summary)
}
\arguments{
\item{stock.summary}{stock.summary from the pre or post season list}

\item{run.year}{inlcude.old.GS.MUs logical indicating if the ML and VI GS MUs should still be in the table}
}
\description{
Combines the GS MUs for the stock.summary DF (pre or post season)
}
