% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{CompilePscData}
\alias{CompilePscData}
\title{Compile a FRAM model run into PSC fishery and stock groupings}
\usage{
CompilePscData(
  fram.db.conn,
  run.name,
  run.year,
  psc.data.list,
  tamm.data.list,
  report.dir,
  combine.GS
)
}
\arguments{
\item{fram.db.conn}{RODBC connection to FRAM access database}

\item{run.name}{FRAM run name that is to be compiled into PSC fisheries/stocks}

\item{psc.data.list}{Data list of FRAM to PSC stocks and fisheries}

\item{tamm.data.list}{Data list related to TAMM Excel Workbook interpretation}
}
\value{
A data frame with PSC fishery mortalities by PSC Stock
}
\description{
Compile a FRAM model run into PSC fishery and stock groupings
}
