% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImportFiles.R
\name{createImportFiles}
\alias{createImportFiles}
\title{Create Import Files}
\usage{
createImportFiles(
  fram.run.name,
  fram.db.name = NA,
  report.dir = "./import_files/",
  data.dir = "./csv/",
  connection_driver = "Driver={Microsoft Access Driver (*.mdb, *.accdb)};"
)
}
\arguments{
\item{fram.run.name}{Preseason run name to use to}

\item{fram.db.name}{.mdb file that contains the preason runs}

\item{report.dir}{the path where the import files are written to}

\item{data.dir}{path where the .csv files for the package are located}
}
\description{
Create Import Files
}
