% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{CreateTable2}
\alias{CreateTable2}
\title{Generate the Table 2 of the PSC Annual report for the specified run year}
\usage{
CreateTable2(pre.season.data, post.season.data, run.year)
}
\arguments{
\item{pre.season.data}{PSC summarized data by fishery and stock from Pre-Season model run}

\item{post.season.data}{PSC summarized data by fishery and stock from Post-Season model run}

\item{run.year}{The run year that table is generated for}
}
\value{
Table 2 of the annual report as a data frame
}
\description{
Generate the Table 2 of the PSC Annual report for the specified run year
}
