% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportERs.R
\name{exportFisheryERs}
\alias{exportFisheryERs}
\title{xlsx of fishery ers}
\usage{
exportFisheryERs(
  fram.db = NA,
  run.names,
  report.dir = "./report/",
  connection_driver = "Driver={Microsoft Access Driver (*.mdb, *.accdb)};"
)
}
\arguments{
\item{fram_db}{Data base with the FRAM runs}

\item{run_names}{Vector of run names from the FRAM database}
}
\description{
xlsx of fishery ers
}
