% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\docType{data}
\name{FisheryMortSqlFilename}
\alias{FisheryMortSqlFilename}
\title{Constants for dealing with a FRAM Database}
\format{
An object of class \code{character} of length 1.
}
\usage{
FisheryMortSqlFilename
}
\description{
Constants for dealing with a FRAM Database
}
\keyword{datasets}
