% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{FormatDouble}
\alias{FormatDouble}
\title{Formats a value to an double in text form with the specified number of decimal places.
This is typically used to generate pretty report text values
 @param  value A vector of values to convert to Double format. Can be factor, character, or number
 @param decimals how many decimal places the final values should display}
\usage{
FormatDouble(value, decimals)
}
\value{
vector of text values in integer format (i.e. number with no decimal place).
}
\description{
Formats a value to an double in text form with the specified number of decimal places.
This is typically used to generate pretty report text values
 @param  value A vector of values to convert to Double format. Can be factor, character, or number
 @param decimals how many decimal places the final values should display
}
