% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatText.R
\name{formatHtml}
\alias{formatHtml}
\title{Retreives HTML text, binds parameters to the text}
\usage{
formatHtml(
  html_file_name,
  params = NULL,
  html_dir = "html_templates",
  package_name = packageName(parent.frame())
)
}
\arguments{
\item{html_file_name}{The name of of the file with HTML to load}

\item{params}{List of named parameters to bind to the HTML text}

\item{html_dir}{The directory that stores all the HTML tempalte documents (defaults to "html")}

\item{package_name}{The package name to load sql file from, defaults to the current package name}
}
\value{
An html text statement
}
\description{
Retreives HTML text, binds parameters to the text
}
