% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{FormatInt}
\alias{FormatInt}
\title{Formats a value to an integer in text form.
This is typically used to generate pretty report text values}
\usage{
FormatInt(values)
}
\arguments{
\item{values}{A vector of values to convert to Integer format. Can be factor, character, or number}
}
\value{
A vector of text values in integer format (i.e. number with no decimal place).
}
\description{
Formats a value to an integer in text form.
This is typically used to generate pretty report text values
}
