% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{FormatNameIdText}
\alias{FormatNameIdText}
\title{Formats parallel vectors of names and IDs a single text block.
This is mainly used to identify fisheries or stocks with a particular issue.}
\usage{
FormatNameIdText(names, ids, show_total = 5, sep = "\\n")
}
\arguments{
\item{names}{A vector of names to format}

\item{ids}{A vector of ids to format}

\item{show_total}{How many lines of text_lines to show in the output}

\item{sep}{Character used to seperate the lines}
}
\value{
A single text block.
}
\description{
Formats parallel vectors of names and IDs a single text block.
This is mainly used to identify fisheries or stocks with a particular issue.
}
