% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatText.R
\name{formatPkgTextFile}
\alias{formatPkgTextFile}
\title{Get the path to a file within an R package}
\usage{
formatPkgTextFile(
  file_name,
  params = NULL,
  dir,
  clean_sql = TRUE,
  package_name = packageName(parent.frame())
)
}
\arguments{
\item{file_name}{File name to load from the R package}

\item{params}{List of Parameter names use to replace variables in the file using str_interp (optional)}

\item{dir}{The schema name where the data is stored (e.g MRP_OPS)}

\item{clean_sql}{Should the format of SQL text be cleaned up (e.g. replace newline characters with a space)}

\item{package_name}{Package name that the file is located in.}
}
\value{
Text from the file name provided
}
\description{
Get the path to a file within an R package
}
