% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatText.R
\name{formatSql}
\alias{formatSql}
\title{Retreives Sql Statement text from a file, binds parameters to the text, and  do basic format cleanup}
\usage{
formatSql(
  sql_file_name,
  params = NULL,
  sql_dir = "sql",
  clean_sql = TRUE,
  package_name = packageName(parent.frame())
)
}
\arguments{
\item{sql_file_name}{The name of of the file with sql to load}

\item{params}{List of named parameters to bind to the sql text}

\item{sql_dir}{The directory that stores all the sql statements (defaults to "sql")}

\item{clean_sql}{A boolean that identifies if the format of the final sql should be cleaned (e.g. strip newlines)}

\item{package_name}{The package name to load sql file from, defaults to the current package name}
}
\value{
An Sql text statement
}
\description{
Retreives Sql Statement text from a file, binds parameters to the text, and  do basic format cleanup
}
