% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\docType{data}
\name{FramGetSingleRecruitScalar}
\alias{FramGetSingleRecruitScalar}
\title{Recruit Scalar FRAM SQL scripts  ---------------------------------------------------------}
\format{
An object of class \code{character} of length 1.
}
\usage{
FramGetSingleRecruitScalar
}
\description{
Recruit Scalar FRAM SQL scripts  ---------------------------------------------------------
}
\keyword{datasets}
