% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{GetColumnSums}
\alias{GetColumnSums}
\title{Sum columns}
\usage{
GetColumnSums(data, label.column, label.value)
}
\arguments{
\item{data}{data frame}

\item{label.column}{labels}

\item{label.value}{value}
}
\description{
Sum columns
}
