% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramBackwardEscapement}
\alias{getFramBackwardEscapement}
\title{A helper function retrieving the escapement values used by the backward FRAM during post-season run}
\usage{
getFramBackwardEscapement(fram.db.conn, fram.run.name)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{run.name}{The name of the model run you would like to retrieve backward FRAM Escapement values for}
}
\value{
A data frame with the Backward FRAM escapement data, based on the model run name provided
}
\description{
A helper function retrieving the escapement values used by the backward FRAM during post-season run
}
