% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramBaseStocks}
\alias{getFramBaseStocks}
\title{Get the data frame of valid stocks from the base period of a specific model run}
\usage{
getFramBaseStocks(fram.db.conn, fram.run.name)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{fram.run.name}{The name of the model run you would like to retrieve fisheries and timesteps from}
}
\value{
A dataframe with the fishery scalars for a specific model run name
}
\description{
Get the data frame of valid stocks from the base period of a specific model run
}
