% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramData}
\alias{getFramData}
\title{Generic function that retrieves data from database using sql file}
\usage{
getFramData(
  db_conn,
  sql_file_name,
  params = NULL,
  sql_dir = "sql",
  clean_sql = TRUE,
  package_name = packageName(parent.frame())
)
}
\arguments{
\item{db_conn}{A connection to the MRP Operational database}

\item{params}{Parameters to insert into the SQL statement}

\item{sql_filename}{A text file containing SQL file}
}
\description{
Generic function that retrieves data from database using sql file
}
