% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramFisheries}
\alias{getFramFisheries}
\title{A helper function loading the list of FRAM fisheries in the database}
\usage{
getFramFisheries(fram.db.conn)
}
\arguments{
\item{fram.db.conn}{An odbc connection to the FRAM database}
}
\value{
A data frame of FRAM fisheries
}
\description{
A helper function loading the list of FRAM fisheries in the database
}
