% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramFisheryMortality}
\alias{getFramFisheryMortality}
\title{A helper function loading the total mortalities for all fisheries and time steps within a FRAM model run}
\usage{
getFramFisheryMortality(fram.db.conn, run.name, run.year)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{run.name}{The name of the model run you would like to load fishery mortalities for}
}
\value{
A dataframe with the mortalities from the FRAM fisheries and time steps for a specific model run name
}
\description{
A helper function loading the total mortalities for all fisheries and time steps within a FRAM model run
}
\section{Exceptions}{

  The method checks the run year of the model run against a provided value, if they don't match
  then the method throws an exception.
}

