% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramFisheryScalars}
\alias{getFramFisheryScalars}
\title{Get the dataframe of fishery scalars used to parameterize model runs}
\usage{
getFramFisheryScalars(fram.db.conn, fram.run.name)
}
\arguments{
\item{fram.db.conn}{An odbc connection to the FRAM database}

\item{run.name}{The name of the model run you would like to retrive fishery scalars from}
}
\value{
A dataframe with the fishery scalars for a specific model run name
}
\description{
Get the dataframe of fishery scalars used to parameterize model runs
}
