% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramRunInfo}
\alias{getFramRunInfo}
\title{Retrieve the details about a specific FRAM run, by run name}
\usage{
getFramRunInfo(fram.db.conn, fram.run.name)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{fram.run.name}{The FRAM run name that details are requested for}
}
\value{
FRAM run details in a data frame
}
\description{
Retrieve the details about a specific FRAM run, by run name
}
