% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramRunTable}
\alias{getFramRunTable}
\title{Retrieve all FRAM runs in a specific data base and species}
\usage{
getFramRunTable(fram.db.conn, species.name = kCohoSpeciesName)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{species.name}{The species for the query}
}
\value{
FRAM run details in a data frame
}
\description{
Retrieve all FRAM runs in a specific data base and species
}
