% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramStockRecruitScalars}
\alias{getFramStockRecruitScalars}
\title{Get the data frame of stock recruit scalars for particular model runs}
\usage{
getFramStockRecruitScalars(fram.db.conn, fram.run.name)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{fram.run.name}{The name of the model run you would like to retrieve fisheries and timesteps from}
}
\value{
A data frame with the stock recruit scalars for a specific model run name
}
\description{
Get the data frame of stock recruit scalars for particular model runs
}
