% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{getFramStocks}
\alias{getFramStocks}
\title{A helper function loading the list of FRAM stocks}
\usage{
getFramStocks(fram.db.conn)
}
\arguments{
\item{fram.db.conn}{An odbc connection to the FRAM database}
}
\value{
A data fram of FRAM stocks
}
\description{
A helper function loading the list of FRAM stocks
}
