% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PscFramAdminData.r
\name{getPersonFramFisheries}
\alias{getPersonFramFisheries}
\title{Retrieve a data frame with the relationships between CoTC members and FRAM fisheries}
\usage{
getPersonFramFisheries(select_person_name = NA, data_dir = data.dir)
}
\arguments{
\item{select_person_name}{Provide a person name to filter results or NA if all person fisheries}

\item{data_dir}{The directory that the file with the person fishery definition file is in}
}
\value{
A data frame identifying who is resposible for each FRAM fishery catch reporting
}
\description{
Retrieve a data frame with the relationships between CoTC members and FRAM fisheries
}
