% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preseasonExport.R
\name{getPreseasonERs}
\alias{getPreseasonERs}
\title{Writes of preseaon FRAM data}
\usage{
getPreseasonERs(
  run.year,
  pre.season.run.name,
  pre.season.fram.db = NA,
  pre.season.tamm = NA,
  pre.season.tamm.fishery.ref = "./csv/TammFisheryFullRef.csv",
  pre.season.tamm.esc.ref = "./csv/TammEscFullRef.csv",
  template.dir = "./templates/",
  report.dir = "./report/",
  data.dir = "./csv/",
  combine.GS = TRUE,
  big.bar.esc = 0,
  big.bar.morts = 0,
  connection_driver = "Driver={Microsoft Access Driver (*.mdb, *.accdb)};"
)
}
\arguments{
\item{run.year}{The fishery year for the model run}

\item{pre.season.run.name}{Run name of the pre season run fron the RunID table}

\item{pre.season.fram.db}{Fram DB containing the pre season run.  NA opens file selection}

\item{pre.season.tamm}{Tamm excel for from pre season run.  NA opens file selection}

\item{pre.season.tamm.fishery.ref}{Fishery references for pre season TAMM file}

\item{pre.season.tamm.esc.ref}{Escapement references for pre season TAMM file}

\item{template.dir}{Template folder}

\item{report.dir}{directory to write reports to}

\item{data.dir}{folder containing the reference .csv files}

\item{combine.GS}{logical if GS MUs should be combined}
}
\description{
Writes of preseaon FRAM data
}
