% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{GetPscEscapement}
\alias{GetPscEscapement}
\title{Sum up the escapement from a fram run to PSC Stocks}
\usage{
GetPscEscapement(escapement, psc.stock, psc.stock.map)
}
\arguments{
\item{escapement}{Escapement values from a FRAM run}

\item{psc.stock}{The list of PSC stocks}

\item{psc.stock.map}{Relates FRAM stocks to PSC stocks}
}
\value{
A data frame with PSC escapement
}
\description{
Sum up the escapement from a fram run to PSC Stocks
}
