% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{GetPscMortality}
\alias{GetPscMortality}
\title{Sum up the fishery mortalities from a fram run to PSC Stocks}
\usage{
GetPscMortality(
  fishery.mortality,
  psc.fishery,
  psc.fishery.map,
  psc.stock,
  psc.stock.map
)
}
\arguments{
\item{fishery.mortality}{Fishery mortality values from a FRAM run}

\item{psc.fishery}{The list of PSC fisheries}

\item{psc.fishery.map}{Relates FRAM fisheries to PSC fisheries}

\item{psc.stock}{The list of PSC stocks}

\item{psc.stock.map}{Relates FRAM stocks to PSC stocks}
}
\value{
A data frame with PSC fishery mortalities by PSC Stock
}
\description{
Sum up the fishery mortalities from a fram run to PSC Stocks
}
