% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{GetPstStockCountryCap}
\alias{GetPstStockCountryCap}
\title{Calculate Country Specific PSC Stock ER Cap based on stock status and the PST}
\usage{
GetPstStockCountryCap(stock.status)
}
\arguments{
\item{stock.status:}{Post Season data in PSC fishery/stock format}
}
\value{
The stock.status data frame with the US and Canadian ER Caps
}
\description{
Calculate Country Specific PSC Stock ER Cap based on stock status and the PST
}
