% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{GetPstStockStatusCap}
\alias{GetPstStockStatusCap}
\title{Identifies the stock status of PSC Stocks based on the Pacific Salmon Treaty}
\usage{
GetPstStockStatusCap(stock.summary, run.year)
}
\arguments{
\item{stock.summary}{Summary of stock information for PSC Stocks}

\item{run.year}{The run year (used for a kludge to handle 2014 IFR Coho status)}
}
\value{
The stock.status data frame with the US and Canadian ER Caps
}
\description{
Identifies the stock status of PSC Stocks based on the Pacific Salmon Treaty
}
