% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TammData.r
\name{GetTammData}
\alias{GetTammData}
\title{Reads the various values of the TAMM spreadsheet and packages data into a list}
\usage{
GetTammData(tamm.filename, tamm.fishery.ref.filename, tamm.esc.ref.filename)
}
\arguments{
\item{tamm.filename}{The file name of TAMM excel spreadsheet}

\item{data.dir}{Directory that TAMM reference files are defined}
}
\value{
A list with a dataframe for fishery mortalities and escapement values
}
\description{
Reads the various values of the TAMM spreadsheet and packages data into a list
}
