% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TammData.r
\name{GetTammEscapement}
\alias{GetTammEscapement}
\title{Reads the specific Escapement values from a TAMM model defined in the
tamm.esc.ref.filename}
\usage{
GetTammEscapement(tamm.filename, tamm.esc.ref.filename)
}
\arguments{
\item{tamm.filename}{The file name of TAMM excel spreadsheet}

\item{tamm.esc.ref.filename}{The file name of TAMM excel spreadsheet}
}
\value{
A dataframe with the FRAM stock ID and associated TAMM escapement
}
\description{
Reads the specific Escapement values from a TAMM model defined in the
tamm.esc.ref.filename
}
\note{
The method checks that values read for the Excel spreadsheet are numeric values.
}
