% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TammData.r
\name{GetTammFisheryMortality}
\alias{GetTammFisheryMortality}
\title{Reads the specific Stock/Fishery mortality values from a TAMM model defined in the
provided file.}
\usage{
GetTammFisheryMortality(tamm.filename, tamm.fishery.ref.filename)
}
\arguments{
\item{tamm.filename}{The file name of TAMM excel spreadsheet}

\item{tamm.fishery.ref.filename}{The file name containing}
}
\value{
A dataframe with the FRAM fisheries/stock combination and associated TAMM mortalties
}
\description{
Reads the specific Stock/Fishery mortality values from a TAMM model defined in the
provided file.
}
\note{
The method checks that values read for the Excel spreadsheet are numeric values.
}
