% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TammData.r
\name{GetTammValues}
\alias{GetTammValues}
\title{A helper function loading the total mortalities from the TAMM excel spreadsheet
based on a the worksheet name, column, and row of the cell value defined in the tamm.ref
data frame}
\usage{
GetTammValues(tamm.ref, tamm.filename)
}
\arguments{
\item{tamm.ref}{A data frame of cell references to load from the TAMM Excel document}

\item{tamm.filename:}{The file name of the TAMM Excel document}
}
\value{
A dataframe with the FRAM fisheries and associated TAMM mortalties
}
\description{
A helper function loading the total mortalities from the TAMM excel spreadsheet
based on a the worksheet name, column, and row of the cell value defined in the tamm.ref
data frame
}
