% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{GetTimeStampText}
\alias{GetTimeStampText}
\title{Provides a standardized text based time stamp for inclusion in file names}
\usage{
GetTimeStampText()
}
\value{
A string with the current date and time
}
\description{
Provides a standardized text based time stamp for inclusion in file names
}
