% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{importFramTemplates}
\alias{importFramTemplates}
\title{Import an updated .csv postseason template into a FRAM database}
\usage{
importFramTemplates(
  template_file = NA,
  validate.catch = TRUE,
  validate.fisheries = TRUE,
  validate.mark.info = TRUE,
  validate.stocks = TRUE,
  validate.escapment.flags = TRUE,
  data.dir = "./csv/"
)
}
\arguments{
\item{db_conn}{A connection to the MRP Operational database}

\item{sql_filename}{A text file containing SQL file}

\item{params}{Parameters to insert into the SQL statement}
}
\description{
Import an updated .csv postseason template into a FRAM database
}
