% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\docType{data}
\name{kFramBackwardEscSqlFilename}
\alias{kFramBackwardEscSqlFilename}
\title{Backward FRAM SQL scripts  ---------------------------------------------------------}
\format{
An object of class \code{character} of length 1.
}
\usage{
kFramBackwardEscSqlFilename
}
\description{
Backward FRAM SQL scripts  ---------------------------------------------------------
}
\keyword{datasets}
