% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\docType{data}
\name{kFramGetSingleNonRetention}
\alias{kFramGetSingleNonRetention}
\title{Non-Retention FRAM SQL scripts  ---------------------------------------------------------}
\format{
An object of class \code{character} of length 1.
}
\usage{
kFramGetSingleNonRetention
}
\description{
Non-Retention FRAM SQL scripts  ---------------------------------------------------------
}
\keyword{datasets}
