% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{LoadConfigFiles}
\alias{LoadConfigFiles}
\title{Loads a R source file that is used for configuration}
\usage{
LoadConfigFiles(report.config.file)
}
\arguments{
\item{report.config.file}{The file name that is used to load configuration parameters}
}
\description{
Loads a R source file that is used for configuration
}
