% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{loadPscData}
\alias{loadPscData}
\title{A function that loads the PSC stock and fishery reference tables from CSV files.
The resulting tables are combined into a list}
\usage{
loadPscData(data.dir)
}
\arguments{
\item{data.dir}{The directory where there reference csv files are saved}
}
\value{
A list with the psc.stock, psc.stock.map, psc.fishery, and psc.fishery.map dataframes
}
\description{
A function that loads the PSC stock and fishery reference tables from CSV files.
The resulting tables are combined into a list
}
