% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{LoadSourceFile}
\alias{LoadSourceFile}
\title{Loads an R file.  This function makes sure a given source file is accidently loaded multiple times}
\usage{
LoadSourceFile(source.file.name)
}
\arguments{
\item{source.file.name}{The file name of the R source to be loaded}
}
\value{
A string with the current date and time
}
\description{
Loads an R file.  This function makes sure a given source file is accidently loaded multiple times
}
