% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{pairEscapementFlags}
\alias{pairEscapementFlags}
\title{Matches escapment flags for marked and unmarked stocks, to validate escapement flags
and identify if stock recruit scalars need to be updated.}
\usage{
pairEscapementFlags(target_escapement)
}
\arguments{
\item{target_escapement}{A data frame of the targent escapements from an import file}
}
\value{
The target_escapement with an additional column identify marked/unmarkd combined flag
}
\description{
Matches escapment flags for marked and unmarked stocks, to validate escapement flags
and identify if stock recruit scalars need to be updated.
}
