% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{parseImportFile}
\alias{parseImportFile}
\title{Parses the import file format and returns the results into a list
The file format generally consists of a header followed by a CSV table of fishery catch.
Each section of the file is seperated by 4 or more dashes (e.g. ----------) and a carrage return.}
\usage{
parseImportFile(import.file.name)
}
\arguments{
\item{import.file.name}{The file name of the import file}
}
\value{
A list with the different sections of the import file
}
\description{
Parses the import file format and returns the results into a list
The file format generally consists of a header followed by a CSV table of fishery catch.
Each section of the file is seperated by 4 or more dashes (e.g. ----------) and a carrage return.
}
