% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportERs.R
\name{postSeasonTable3Summary}
\alias{postSeasonTable3Summary}
\title{Creates a .xlsx workbook of the "Table3" from the annual report (post season ERs)}
\usage{
postSeasonTable3Summary(
  fram.db = NA,
  run.names,
  combine.GS = TRUE,
  data.dir = "./csv/"
)
}
\arguments{
\item{fram_db}{Data base with the FRAM runs}

\item{run_names}{Vector of run names from the FRAM database}
}
\description{
Creates a .xlsx workbook of the "Table3" from the annual report (post season ERs)
}
