% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{ReadCsv}
\alias{ReadCsv}
\title{Read a CSV file from a predefined data directory
Optionally, the function can check the file for uniqueness by providing a vector
of column names that should be unique to each row.}
\usage{
ReadCsv(file.name, data.dir, unique.col.names = NULL)
}
\arguments{
\item{file.name}{The file name of the CSV file to load}

\item{data.dir}{The directory that the CSV file is saved}

\item{unique.col.names}{(optional): A vector of column names that combined uniquely identify each row}
}
\description{
This function always assumes that the CSV file has a header row.
}
