% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{ReadMemoryCsv}
\alias{ReadMemoryCsv}
\title{A helper function for reading CSV formated text from a R character variable}
\usage{
ReadMemoryCsv(x)
}
\arguments{
\item{x}{A data frame to be written to as CSV format to a vector of characters}
}
\value{
A character vector with the lines of CSV
}
\description{
A helper function for reading CSV formated text from a R character variable
}
