% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{RunSqlFile}
\alias{RunSqlFile}
\title{A helper function that loads an SQL script, updates the variables in the script to values provide and
formats the resulting data by renames columns to common R style.}
\usage{
RunSqlFile(db.conn, file.name, variables = NA)
}
\arguments{
\item{db.conn}{An ODBC connection to the ODBC database}

\item{file.name}{A file name that the SQL script is saved to}

\item{variables}{An R list of variables, variable names in the list are matched to ones with the same name in
a format like %VARIABLENAME% (eg list(runid = 1) will replace %RUNID% in the SQL with 1)}
}
\value{
A data frame with query results
}
\description{
A helper function that loads an SQL script, updates the variables in the script to values provide and
formats the resulting data by renames columns to common R style.
}
