% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{SummaryRegionTotals}
\alias{SummaryRegionTotals}
\title{Make a table of fishery mortality}
\usage{
SummaryRegionTotals(psc.fishery.table)
}
\arguments{
\item{psc.fishery.table}{Table of fisheries}
}
\value{
Summarize regional totals
}
\description{
Make a table of fishery mortality
}
