% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{TitleCase}
\alias{TitleCase}
\title{Convert text into title case
Got this regex from: http://stackoverflow.com/questions/15776732/how-to-convert-a-vector-of-strings-to-title-case
This function could use \\E (stop capitalization) rather than \\L (start lowercase), depending on what rules you want to follow}
\usage{
TitleCase(text)
}
\arguments{
\item{text}{A vector of strings that are to be converted to title case}
}
\value{
The provided text in title case
}
\description{
Convert text into title case
Got this regex from: http://stackoverflow.com/questions/15776732/how-to-convert-a-vector-of-strings-to-title-case
This function could use \\E (stop capitalization) rather than \\L (start lowercase), depending on what rules you want to follow
}
