% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{updateFisheryScalars}
\alias{updateFisheryScalars}
\title{Update the fishery scalars and non retention values for an identified model run based on
values in a dataframe.  The Non-Retention CNR mortalities updates more intellegently (e.g.
remove/adding/updating DB rows based on values provided and values within the database run)}
\usage{
updateFisheryScalars(fram.db.conn, fram.run.id, fishery.scalars)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{fram.run.id}{The ID of the FRAM model run to update fishery scalars for}

\item{fishery.scalars}{The name of the model run you would like to retrive fishery scalars from}
}
\description{
Update the fishery scalars and non retention values for an identified model run based on
values in a dataframe.  The Non-Retention CNR mortalities updates more intellegently (e.g.
remove/adding/updating DB rows based on values provided and values within the database run)
}
