% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramDB.R
\name{updateTargetEscapement}
\alias{updateTargetEscapement}
\title{Update the stock recruit scalars and the backward FRAM target escapement values/flags.
The target escapement updates more intellegently (e.g. remove/adding/updating DB rows based on
flag and values provided and values within the database run)}
\usage{
updateTargetEscapement(fram.db.conn, fram.run.id, escapement_df)
}
\arguments{
\item{fram.db.conn}{An ODBC connection to the FRAM database}

\item{fram.run.id}{The ID of the FRAM model run to update fishery scalars for}

\item{fishery.scalars}{The name of the model run you would like to retrive fishery scalars from}
}
\description{
Update the stock recruit scalars and the backward FRAM target escapement values/flags.
The target escapement updates more intellegently (e.g. remove/adding/updating DB rows based on
flag and values provided and values within the database run)
}
