% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{ValidateRunInfo}
\alias{ValidateRunInfo}
\title{Checks run data retrieved from the FRAM database.}
\usage{
ValidateRunInfo(run.info, run.year)
}
\arguments{
\item{run.info}{List of model run data retreived from FRAM database}

\item{run.year}{The run year that the run.info should represent}
}
\description{
Checks run data retrieved from the FRAM database.
}
