% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{ValidateValueDomain}
\alias{ValidateValueDomain}
\title{A helper function for validating data within a pre-defined set of possible values.}
\usage{
ValidateValueDomain(
  values,
  domain,
  error.message = "The following values are invalid:\\n\\n\%s\\n\\n"
)
}
\arguments{
\item{values}{The data that is to be validated}

\item{domain}{The set of valid values that "values" can take on.}

\item{error.message}{A template of a printf message for invalid values.}
}
\description{
A helper function for validating data within a pre-defined set of possible values.
}
