% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{validEscapementFlags}
\alias{validEscapementFlags}
\title{Validates the escapement flags against the escapement values for parametrization
of the backward FRAM model run.}
\usage{
validEscapementFlags(target_escapement)
}
\arguments{
\item{target_escapement}{A data frame of the targent escapements from an import file}
}
\value{
A boolean, TRUE for valid or FALSE for when there is issues with the escapement flags
}
\description{
Validates the escapement flags against the escapement values for parametrization
of the backward FRAM model run.
}
