% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{validMarkInfo}
\alias{validMarkInfo}
\title{Validates the mark rate information for mark-selective fisheries.}
\usage{
validMarkInfo(fishery.scalars)
}
\arguments{
\item{fishery.scalars}{The catch data loaded from a post season import file.}
}
\value{
A boolean, TRUE for valid or FALSE for when there is issues with the mark rate information
  with the mark-selective fisheries
}
\description{
Validates the mark rate information for mark-selective fisheries.
}
