% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{validPostSeasonCatch}
\alias{validPostSeasonCatch}
\title{Validates the catch data for parametrization of a FRAM post-season model run.
Most of the validation is related to the appropriate setting of flags and
providing all the necessary parameters.}
\usage{
validPostSeasonCatch(fishery.scalars)
}
\arguments{
\item{fishery.scalars}{The catch data loaded from a post season import file.}
}
\value{
A boolean, TRUE for valid or FALSE for when there is issues with the catch
}
\description{
Validates the catch data for parametrization of a FRAM post-season model run.
Most of the validation is related to the appropriate setting of flags and
providing all the necessary parameters.
}
