% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPostSeasonTemplates.R
\name{validTargetEscapement}
\alias{validTargetEscapement}
\title{Validates the target escapement definitions for parametrization of the FRAM model.
This function checks that all the stocks are valid, relative to the base period
and that all the stocks identified are the responsibility of the identified person.}
\usage{
validTargetEscapement(
  person_name,
  fram_db_conn,
  fram_run_name,
  target_escapement,
  data.dir
)
}
\arguments{
\item{person_name}{the name of the person for the import file}

\item{fram_db_conn}{FRAM ODBC database connection}

\item{fram_run_name}{FRAM run name from the import file}

\item{target_escapement}{Target escapement for a Backward FRAM model run}
}
\value{
A boolean, TRUE for valid or FALSE for when there is issues with the escapement
}
\description{
Validates the target escapement definitions for parametrization of the FRAM model.
This function checks that all the stocks are valid, relative to the base period
and that all the stocks identified are the responsibility of the identified person.
}
