% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePostSeasonReport.R
\name{writeAnnualReport}
\alias{writeAnnualReport}
\title{Writes .pdf document of the annual report}
\usage{
writeAnnualReport(
  run.year,
  post.season.run.name,
  pre.season.run.name,
  post.season.fram.db = NA,
  post.season.tamm = NA,
  post.season.tamm.fishery.ref = "./csv/TammFisheryQueetsRef.csv",
  post.season.tamm.esc.ref = "./csv/TammEscQueetsRef.csv",
  pre.season.fram.db = NA,
  pre.season.tamm = NA,
  pre.season.tamm.fishery.ref = "./csv/TammFisheryFullRef.csv",
  pre.season.tamm.esc.ref = "./csv/TammEscFullRef.csv",
  template.dir = "./templates/",
  report.dir = "./report/",
  data.dir = "./csv/",
  combine.GS = NA,
  big.bar.esc = 39434,
  big.bar.morts = 4392
)
}
\arguments{
\item{run.year}{The fishery year for the model run}

\item{post.season.run.name}{Run name of the post season run fron the RunID table}

\item{pre.season.run.name}{Run name of the pre season run fron the RunID table}

\item{post.season.fram.db}{Fram DB containing the post season run.  Deafaults to NA which prompts a selection to dialog}

\item{post.season.tamm}{Tamm excel file containing the post season data. Deafaults to NA which prompts a selection to dialog}

\item{post.season.tamm.fishery.ref}{Fishery references for post season TAMM file}

\item{post.season.tamm.esc.ref}{Escapement reference of post season TAMM}

\item{pre.season.fram.db}{Fram DB containing the pre season run.  NA opens file selection}

\item{pre.season.tamm}{Tamm excel for from pre season run.  NA opens file selection}

\item{pre.season.tamm.fishery.ref}{Fishery references for pre season TAMM file}

\item{pre.season.tamm.esc.ref}{Escapement references for pre season TAMM file}

\item{template.dir}{Template folder}

\item{report.dir}{directory to write reports to}

\item{data.dir}{folder containing the reference .csv files}
}
\description{
Writes .pdf document of the annual report
}
