% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{WriteCsv}
\alias{WriteCsv}
\title{A helper function for writing CSV files. This is a simple wrapper to handle default flags in other script files}
\usage{
WriteCsv(file, data)
}
\arguments{
\item{file}{The .csv file}

\item{data}{file to write to}
}
\description{
A helper function for writing CSV files. This is a simple wrapper to handle default flags in other script files
}
