% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImportFiles.R
\name{writeImportFile}
\alias{writeImportFile}
\title{Functions for writing .csv files for CoTC scientists to update and then import to FRAM db
used in a loop in creatImportFiles fuction}
\usage{
writeImportFile(
  person_name,
  fram_run_name,
  fram_run_id,
  fram_db_name,
  person_fishery_scalars,
  person_escapement
)
}
\arguments{
\item{person_name}{the individual scientist to whom the file is sent.  USsd for naming}

\item{fram_run_name}{the fram run name used to creat the files}

\item{fram_run_id}{from run number used to create the files}

\item{fram_db_name}{FRAM database used to create the files database}

\item{person_fishery_scalars}{dataframe of scalars}

\item{person_escapement}{dataframe of escapement}
}
\description{
Functions for writing .csv files for CoTC scientists to update and then import to FRAM db
used in a loop in creatImportFiles fuction
}
