% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{WriteMemoryCsv}
\alias{WriteMemoryCsv}
\title{A helper function for writing CSV file format text into R vector of characters}
\usage{
WriteMemoryCsv(x)
}
\arguments{
\item{x}{A data frame to be written to as CSV format to a vector of characters}
}
\value{
A character vector with the lines of CSV
}
\description{
A helper function for writing CSV file format text into R vector of characters
}
