% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.r
\name{WriteProtectedCsv}
\alias{WriteProtectedCsv}
\title{A helper function for writing CSV files and setting them to readonly for protection.
When writing the file, the file is first set to writable then written over then set to readonly}
\usage{
WriteProtectedCsv(data, file.name)
}
\arguments{
\item{data}{Data frame to be written to CSV file}

\item{file.name}{File name to write CSV file to}
}
\description{
A helper function for writing CSV files and setting them to readonly for protection.
When writing the file, the file is first set to writable then written over then set to readonly
}
