% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualReportLib.r
\name{WritePSCFramTables}
\alias{WritePSCFramTables}
\title{writes .csv of PSC tables}
\usage{
WritePSCFramTables(
  fram.db.conn,
  psc.fishery,
  psc.fishery.map,
  psc.stock,
  psc.stock.map,
  report.dir
)
}
\arguments{
\item{fram.db.conn}{connection to fram db}

\item{psc.fishery}{table}

\item{psc.fishery.map}{table}

\item{psc.stock}{table}

\item{psc.stock.map}{table}
}
\value{
Summarize regional totals
}
\description{
writes .csv of PSC tables
}
