% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportERs.R
\name{exportFisheryERs}
\alias{exportFisheryERs}
\title{Writes .pdf document of the annual report}
\usage{
exportFisheryERs(fram.db = NA, run.names, report.dir = "./report/")
}
\arguments{
\item{fram_db}{Data base with the FRAM runs}

\item{run_names}{Vector of run names from the FRAM database}
}
\description{
Writes .pdf document of the annual report
}
